+++ src/ll/ul4c.py
@@ -201,16 +201,25 @@ class UndefinedVariable(Undefined):
 class Context:
 	"""
 	A :class:`Context` object stores the context of a call to a template. This
-	consists of the local variables and the indent stack.
+	consists of local, global and builtin variables and the indent stack.
 	"""
-	# "Global" functions. Will be exposed to UL4 code
+	# "Builtin" functions. Will be exposed to UL4 code
 	functions = {}
 
-	def __init__(self):
-		self.vars = {}
+	def __init__(self, globals=None):
+		self._globals = globals if globals is not None else {}
+		self.vars = collections.ChainMap({}, self.globals, self.functions)
 		self.indents = [] # Stack of additional indentations for the ``<?render?>`` tag
 		self.escapes = [] # Stack of functions for escaping the output
 		self.asts = [] # Call stack (of :class:`AST` objects)
+
+	@property
+	def globals(self):
+		return self._globals
+
+	@globals.setter
+	def globals(self, vars):
+		self.vars.maps[-2] = self._globals = vars
 
 	@classmethod
 	def makefunction(cls, f):
@@ -222,21 +231,20 @@ class Context:
 
 	@contextlib.contextmanager
 	def replacevars(self, vars):
-		oldvars = self.vars
+		oldvars = self.vars.maps[0]
 		try:
-			self.vars = vars
+			self.vars.maps[0] = vars
 			yield
 		finally:
-			self.vars = oldvars
+			self.vars.maps[0] = oldvars
 
 	@contextlib.contextmanager
 	def chainvars(self):
-		oldvars = self.vars
 		try:
-			self.vars = collections.ChainMap({}, self.vars)
+			self.vars.maps.insert(0, {})
 			yield
 		finally:
-			self.vars = oldvars
+			del self.vars.maps[0]
 
 	def output(self, string):
 		for escape in self.escapes:
@@ -2128,10 +2136,7 @@ class Var(Code):
 		try:
 			return context.vars[self.name]
 		except KeyError:
-			try:
-				return context.functions[self.name]
-			except KeyError:
-				return UndefinedVariable(self.name)
+			return UndefinedVariable(self.name)
 
 	@_handleexpressioneval
 	def evalset(self, context, value):
@@ -3814,12 +3819,16 @@ class Template(Block):
 	of the first ``<?return?>`` tag encountered). In this case all output of the
 	template will be ignored.
 
+	For rendering and calling a template with global variables the following
+	methods are available: :meth:`render_with_globals`,
+	:meth:`renders_with_globals` and :meth:`call_with_globals`.
+
 	A :class:`Template` object is itself an AST node. Evaluating it will store
 	the template object under its name in the local variables.
 	"""
 	ul4attrs = Block.ul4attrs.union({"signature", "doc", "name", "whitespace", "startdelim", "enddelim", "parenttemplate", "fullsource", "renders"})
 
-	version = "48"
+	version = "49"
 
 	output = False # Evaluating a template doesn't produce output, but simply stores it in a local variable
 
@@ -4084,8 +4093,7 @@ class Template(Block):
 	@classmethod
 	def loads(cls, data):
 		"""
-		The class method :meth:`loads` loads the template from string ``data``.
-		``data`` must contain the template in compiled UL4ON format.
+		Loads a template as an UL4ON dump from the string ``data``.
 		"""
 		from ll import ul4on
 		return ul4on.loads(data)
@@ -4093,8 +4101,7 @@ class Template(Block):
 	@classmethod
 	def load(cls, stream):
 		"""
-		The class method :meth:`load` loads the template from the stream
-		``stream``. The stream must contain the template in compiled UL4ON
+		Loads the template as an UL4ON dump from the stream ``stream``.
 		format.
 		"""
 		from ll import ul4on
@@ -4102,22 +4109,21 @@ class Template(Block):
 
 	def dump(self, stream):
 		"""
-		:meth:`dump` dumps the template in compiled UL4ON format to the
-		stream ``stream``.
+		Dump the template in compiled UL4ON format to the stream ``stream``.
 		"""
 		from ll import ul4on
 		ul4on.dump(self, stream)
 
 	def dumps(self):
 		"""
-		:meth:`dumps` returns the template in compiled UL4ON format
-		(as a string).
+		Return the template in compiled UL4ON format (as a string).
 		"""
 		from ll import ul4on
 		return ul4on.dumps(self)
 
 	def _renderbound(self, context):
-		# Helper method used by :meth:`render` and :meth:`TemplateClosure.render` where arguments have already been bound
+		# Helper method used by :meth:`render` and :meth:`TemplateClosure.render`
+		# where arguments have already been bound
 		try:
 			# Bypass ``self.eval()`` which simply stores the object as a local variable
 			# Also bypass ``super().eval()`` as this would add additional stackframe in exception messages
@@ -4140,20 +4146,36 @@ class Template(Block):
 	def render(*args, **kwargs):
 		"""
 		Render the template iteratively (i.e. this is a generator).
-		``args`[1:]` and ``kwargs`` contain the top level variables available
-		to the template code. (``args`[0]` is the ``self`` parameter, but
-		:meth:`render` is defined in this way, to allow a keyword argument named
-		``self``).
+
+		``args[1:]`` and ``kwargs`` contain the top level positional and keyword
+		arguments available to the template code. (``args`[0]` is the ``self``
+		parameter, but :meth:`render` is defined in this way, to allow a keyword
+		argument named ``self``). Positional arguments will only be supported if
+		the template has a signature.
 		"""
 		context = Context()
 		yield from args[0].ul4render(context, *args[1:], **kwargs)
 
+	def render_with_globals(self, args, kwargs, globals):
+		"""
+		Render the template iteratively (i.e. this is a generator).
+
+		``args`` and ``kwargs`` contain the top level positional and keyword
+		arguments available to the template code. ``globals`` contains global
+		variables. Positional arguments will only be supported if the template
+		has a signature.
+		"""
+		context = Context(globals)
+		yield from self.ul4render(context, *args, **kwargs)
+
 	def _rendersbound(self, context):
-		# Helper method used by :meth:`renders` and :meth:`TemplateClosure.renders` where arguments have already been bound
+		# Helper method used by :meth:`renders` and :meth:`TemplateClosure.renders`
+		# where arguments have already been bound
 		return "".join(self._renderbound(context))
 
+	# This will be exposed to UL4 as ``renders``
 	@withcontext
-	def ul4renders(*args, **kwargs): # This will be exposed to UL4 as ``renders``
+	def ul4renders(*args, **kwargs):
 		self = args[0]
 		context = args[1]
 		args = args[2:]
@@ -4163,16 +4185,32 @@ class Template(Block):
 
 	def renders(*args, **kwargs):
 		"""
-		Render the template as a string. ``args`[1:]` and ``kwargs`` contain
-		the top level variables available to the template code. (``args`[0]`
-		is the ``self`` parameter, but :meth:`renders` is defined in this way,
-		to allow a keyword argument named ``self``).
+		Render the template as a string.
+
+		``args[1:]`` and ``kwargs`` contain the top level positional and keyword
+		arguments available to the template code. (``args`[0]` is the ``self``
+		parameter, but :meth:`renders` is defined in this way, to allow a keyword
+		argument named ``self``). Positional arguments will only be supported if
+		the template has a signature.
 		"""
 		context = Context()
 		return args[0].ul4renders(context, *args[1:], **kwargs)
 
+	def renders_with_globals(self, args, kwargs, globals):
+		"""
+		Render the template as a string.
+
+		``args`` and ``kwargs`` contain the top level positional and keyword
+		arguments available to the template code. ``globals`` contains global
+		variables. Positional arguments will only be supported if the template
+		has a signature.
+		"""
+		context = Context(globals)
+		return self.ul4renders(context, *args, **kwargs)
+
 	def _callbound(self, context):
-		# Helper method used by :meth:`__call__` and :meth:`TemplateClosure.__call__` where arguments have already been bound
+		# Helper method used by :meth:`__call__` and :meth:`TemplateClosure.__call__`
+		# where arguments have already been bound
 		try:
 			for output in super().eval(context): # Bypass ``self.eval()`` which simply stores the object as a local variable
 				pass # Ignore all output
@@ -4183,10 +4221,12 @@ class Template(Block):
 	def ul4call(*args, **kwargs):
 		"""
 		Call the template as a function and return the resulting value.
-		``args`[1:]` and ``kwargs`` contain the top level variables available
-		to the template code. (``args`[0]` is the ``self`` parameter, but
-		:meth:`ul4call` is defined in this way, to allow a keyword argument named
-		``self``).
+
+		``args[1:]`` and ``kwargs`` contain the top level positional and keyword
+		arguments available to the template code. (``args`[0]` is the ``self``
+		parameter, but :meth:`ul4call` is defined in this way, to allow a keyword
+		argument named ``self``). Positional arguments will only be supported if
+		the template has a signature.
 		"""
 		self = args[0]
 		context = args[1]
@@ -4198,13 +4238,27 @@ class Template(Block):
 	def __call__(*args, **kwargs):
 		"""
 		Call the template as a function and return the resulting value.
-		``args`[1:]` and ``kwargs`` contain the top level variables available
-		to the template code. (``args`[0]` is the ``self`` parameter, but
-		:meth:`__call__` is defined in this way, to allow a keyword argument named
-		``self``).
+
+		``args[1:]`` and ``kwargs`` contain the top level positional and keyword
+		arguments available to the template code. (``args`[0]` is the ``self``
+		parameter, but :meth:`ul4call` is defined in this way, to allow a keyword
+		argument named ``self``). Positional arguments will only be supported if
+		the template has a signature.
 		"""
 		context = Context()
 		return args[0].ul4call(context, *args[1:], **kwargs)
+
+	def call_with_globals(self, args, kwargs, globals):
+		"""
+		Call the template as a function and return the resulting value.
+
+		``args`` and ``kwargs`` contain the top level positional and keyword
+		arguments available to the template code. ``globals`` contains global
+		variables. Positional arguments will only be supported if the template
+		has a signature.
+		"""
+		context = Context(globals)
+		return self.ul4call(context, *args, **kwargs)
 
 	def jssource(self):
 		"""
@@ -5271,7 +5325,7 @@ class TemplateClosure(Block):
 
 	def __init__(self, template, context, signature):
 		self.template = template
-		self.vars = context.vars
+		self.vars = context.vars.maps[0]
 		self.signature = signature
 
 	@withcontext
@@ -5286,8 +5340,9 @@ class TemplateClosure(Block):
 			# (which wouldn't work anyway as ``self.template.signature`` is an :class:`AST` object)
 			yield from self.template._renderbound(context)
 
+	# This will be exposed to UL4 as ``renders``
 	@withcontext
-	def ul4renders(*args, **kwargs): # This will be exposed to UL4 as ``renders``
+	def ul4renders(*args, **kwargs):
 		self = args[0]
 		context = args[1]
 		args = args[2:]